+++ test/src/my_impl.dart
@@ -1,0 +1,7 @@
+abstract class MyAbstr {
+  String get name;
+}
+
+class MyImpl implements MyAbstr {
+  String name;
+}