+++ test/src/serializer.dart
@@ -5,6 +5,7 @@ import 'package:dartson/transformers/date_time.dart';
 
 import 'my_class.dart';
 import 'sub_class.dart';
+import 'my_impl.dart';
 
 // ignore: uri_has_not_been_generated
 part 'serializer.g.dart';
@@ -15,5 +16,8 @@ part 'serializer.g.dart';
     SubClass,
   ],
   transformers: [DateTimeParser],
+  replacements: {
+    MyAbstr: MyImpl,
+  },
 )
 final Dartson<Map<String, dynamic>> serializer = _serializer$dartson;