+++ test/src/serializer.g.dart
@@ -25,6 +25,7 @@ Map<String, dynamic> _MyClass$encoder(MyClass object, Dartson inst) {
       object.subClasses?.map((e) => _SubClass$encoder(e, inst))?.toList();
   obj['complexMap'] =
       object.complexMap?.map((k, e) => MapEntry(k, _SubClass$encoder(e, inst)));
+  obj['replacement'] = _MyImpl$encoder(object.replacement, inst);
   obj['inherited'] = object.inherited;
   obj['inheritName'] = object.inheritedRenamed;
   return obj;
@@ -50,6 +51,7 @@ MyClass _MyClass$decoder(Map<String, dynamic> data, Dartson inst) {
       ?.toList();
   obj.complexMap = (data['complexMap'] as Map<String, dynamic>)
       ?.map((k, e) => MapEntry(k, _SubClass$decoder(e, inst)));
+  obj.replacement = _MyImpl$decoder(data['replacement'], inst);
   obj.inherited = data['inherited'] as bool;
   obj.inheritedRenamed = data['inheritName'] as String;
   return obj;
@@ -106,13 +108,32 @@ SubClass _SubClass$decoder(Map<String, dynamic> data, Dartson inst) {
   return obj;
 }
 
+Map<String, dynamic> _MyImpl$encoder(MyImpl object, Dartson inst) {
+  if (object == null) {
+    return null;
+  }
+  final obj = new Map<String, dynamic>();
+  obj['name'] = object.name;
+  return obj;
+}
+
+MyImpl _MyImpl$decoder(Map<String, dynamic> data, Dartson inst) {
+  if (data == null) {
+    return null;
+  }
+  final obj = new MyImpl();
+  obj.name = data['name'] as String;
+  return obj;
+}
+
 class _Dartson$impl extends Dartson<Map<String, dynamic>> {
   _Dartson$impl()
       : super(<Type, DartsonEntity>{
           MyClass:
               const DartsonEntity<MyClass>(_MyClass$encoder, _MyClass$decoder),
           SubClass: const DartsonEntity<SubClass>(
-              _SubClass$encoder, _SubClass$decoder)
+              _SubClass$encoder, _SubClass$decoder),
+          MyImpl: const DartsonEntity<MyImpl>(_MyImpl$encoder, _MyImpl$decoder)
         });
 }